feat(persona): the act-question becomes its own module, asked on BOTH turn outcomes (slice 1)
Slice 1 of ROOM-AS-RUNNER. Two changes, one structural and one behavioural, and the
structural one is why the behavioural one was invisible for a session.

BEHAVIOUR — the question is no longer reachable only through a silence.

It lived inside the `SettleStep::Passed` arm, so the act-question was asked ONLY when
she declined to speak. A citizen who ANSWERED someone in the room and also held a
claimed card was never asked whether to work it: talking and working were alternatives.
For a colleague they are not. It is now asked on BOTH outcomes — after a spoken reply
(post-send, so the room hears her at the same latency as before) and after a pass.

That was the THIRD defect in this one conditional, after the two fixed hours earlier:
`!directed` excluded every addressed kickoff — which is how dispatch actuates, so the
actuation path and the work gate were mutually exclusive by construction — and an
`InProgress`-only filter demanded the state that starting work is what produces, which
is circular. Any one of the three alone accounts for `persona.turn.work == 0` in every
live round we have ever measured.

STRUCTURE — and this is the actual lesson.

273 lines, nested SIX levels deep, inside a match arm inside the turn loop of a
4,900-line file. That depth is the bug's habitat: a condition buried that far down
declines invisibly, so five competing hypotheses looked identical from outside and
only a full evening of live bisection separated them. Fixing the conditions without
removing the habitat would have left the next defect just as well hidden.

So the act-question is now `persona/act_question.rs` — one file, one concept, its
contract stated at the top: it is a QUESTION never an instruction (she may still pass),
it never decides WHEN to ask (the caller owns that), and it reports on EVERY path
including every decline. service_loop.rs 4,900 → 4,626; the new module is 334.
Per CLAUDE.md: decompose continuously, a new concept gets its own file.

`cycle` is fetched inside from the same global registry the turn loop uses rather than
threaded through as a borrow, and a missing cycle now emits `decision=no_cycle` instead
of silently doing nothing — the one path that could still have declined in the dark.

36 service_loop tests green. Live proof owed: `persona.turn.work` has still never
fired, and that is what the next round measures.
